validation: isolate provenance capture
